Move `duration` and `move-delete` out of the RSVP `advanced-options`. Show them in the block.
