+++ src/modules/blocks/rsvp/move-delete/style.pcss
@@ -3,7 +3,6 @@
 
 	& > button {
 		font-size: 15px !important;
-		letter-spacing: .38px;
 		line-height: 18px;
 		padding: 0;
 
@@ -18,7 +17,7 @@
 		}
 
 		&:last-child {
-			color: red;
+			color: #da394d;
 		}
 	}
 }
