+++ src/modules/blocks/rsvp/move-delete/template.js
